--- analise_data.py
@@ -1,9 +1,0 @@
-import pandas as pd
-import numpy as np
-import numpy as np
-import matplotlib.pyplot as plt
-
-arquivo = 'output/data.csv'
-usecols = ["indice", "metodo", "tipo", "tamanho", "tempo1", "tempo2", "tempo3", "tempo4", "tempo5", "tempo6", "tempo7", "tempo8", "tempo9", "tempo10"]
-df = pd.read_csv(arquivo, index_col = "indice", usecols = usecols)
-
