--- output/data.csv
@@ -1,21 +1,0 @@
-indice,metodo,tipo,tamanho,tempo1,tempo2,tempo3,tempo4,tempo5,tempo6,tempo7,tempo8,tempo9,tempo10
-0,shellSort,1,10000,0.004000,0.000000,0.004000,0.004000,0.000000,0.002000,0.000000,0.000000,0.004000,0.005000
-1,shellSort,1,50000,0.010000,0.008000,0.009000,0.010000,0.012000,0.012000,0.008000,0.012000,0.012000,0.012000
-2,shellSort,1,100000,0.024000,0.024000,0.024000,0.025000,0.024000,0.024000,0.020000,0.027000,0.024000,0.024000
-3,shellSort,1,500000,0.149000,0.145000,0.153000,0.146000,0.157000,0.145000,0.141000,0.140000,0.154000,0.154000
-4,shellSort,1,1000000,0.342000,0.331000,0.321000,0.292000,0.310000,0.319000,0.316000,0.321000,0.312000,0.287000
-5,shellSort,2,10000,0.000000,0.000000,0.000000,0.000000,0.000000,0.008000,0.000000,0.000000,0.000000,0.000000
-6,shellSort,2,50000,0.004000,0.004000,0.004000,0.004000,0.000000,0.003000,0.003000,0.002000,0.002000,0.001000
-7,shellSort,2,100000,0.004000,0.004000,0.004000,0.005000,0.004000,0.008000,0.004000,0.008000,0.004000,0.004000
-8,shellSort,2,500000,0.049000,0.048000,0.042000,0.041000,0.041000,0.048000,0.042000,0.052000,0.045000,0.048000
-9,shellSort,2,1000000,0.100000,0.106000,0.101000,0.112000,0.113000,0.116000,0.132000,0.122000,0.100000,0.118000
-10,mergeSort,1,10000,0.004000,0.004000,0.004000,0.004000,0.004000,0.004000,0.004000,0.004000,0.004000,0.004000
-11,mergeSort,1,50000,0.021000,0.020000,0.022000,0.020000,0.016000,0.016000,0.018000,0.020000,0.017000,0.020000
-12,mergeSort,1,100000,0.042000,0.036000,0.041000,0.034000,0.036000,0.039000,0.040000,0.032000,0.037000,0.040000
-13,mergeSort,1,500000,0.195000,0.184000,0.181000,0.169000,0.157000,0.164000,0.188000,0.181000,0.171000,0.173000
-14,mergeSort,1,1000000,0.340000,0.355000,0.318000,0.339000,0.361000,0.347000,0.326000,0.313000,0.340000,0.366000
-15,mergeSort,2,10000,0.000000,0.004000,0.000000,0.003000,0.002000,0.004000,0.002000,0.001000,0.000000,0.004000
-16,mergeSort,2,50000,0.012000,0.011000,0.013000,0.012000,0.012000,0.012000,0.012000,0.016000,0.016000,0.017000
-17,mergeSort,2,100000,0.028000,0.028000,0.032000,0.025000,0.024000,0.025000,0.028000,0.025000,0.026000,0.024000
-18,mergeSort,2,500000,0.121000,0.134000,0.137000,0.121000,0.146000,0.126000,0.151000,0.143000,0.137000,0.136000
-19,mergeSort,2,1000000,0.276000,0.251000,0.260000,0.248000,0.275000,0.273000,0.304000,0.270000,0.262000,0.240000
